Add files via upload
This is a GUI of an interface an employee will use to enter information to be recorded to a file. There are 2 files to be written to, one is an employee file that will record any information entered while using the 'Enter Button'. The second file is for 'Star Employees'. this file will only be written to if the entered dollar amount is greater than or equal to 8000. This data will be recorded to the stars file when the 'Evaluate Button' is pressed. If the amount entered is less than 8000, nothing will be recorded when the 'Evaluate Button' is pressed. I tried integrating both functions to the 'Enter Button' but was unable to make it work successfully as I'm still learning the Java language.